pyproject.toml
[tool.poetry]
name = "langgraph-agents"
version = "0.1.0"
description = ""
authors = ["debsouryadatta <debsouryadatta@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
langchain = "^0.3.14"
langchain-core = "^0.3.29"
langchain-community = "^0.3.14"
langgraph = "^0.2.60"
langchain-groq = "^0.2.2"
python-dotenv = "^1.0.1"
ipython = "^8.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
